Remove duplication in the localization files
